Separating List Name editing into own component
